+++ packages/realm-react/CHANGELOG.md
@@ -4,7 +4,34 @@
 * None
 
 ### Enhancements
-* None
+* Added the ability to use an existing Realm instance in `RealmProvider` and `createRealmContext`. ([#6714](https://github.com/realm/realm-js/pull/6714))
+```jsx
+// Using RealmProvider
+import { RealmProvider } from "@realm/react";
+
+const realm = new Realm(...);
+
+function MyApp() {
+  return (
+    <RealmProvider realm={realm}> 
+      ...
+    </RealmProvider>
+  );
+}
+
+// Using createRealmContext
+import { createRealmContext } from "@realm/react";
+
+const realm = new Realm(...);
+const { RealmProvider, useRealm } = createRealmContext(realm);
+
+function MyApp() {
+  return (
+    <RealmProvider> 
+      ...
+    </RealmProvider>
+  );
+```
 
 ### Fixed
 * <How to hit and notice issue? what was the impact?> ([#????](https://github.com/realm/realm-js/issues/????), since v?.?.?)